+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 Changelog
 =========
 
+v1.2.0 (2022-02-25)
+-------------------
+
+- A new functionality: it is now possible to delete an entry
+- Bug correction in del_Element_DLList (delete a node at position x of the list)
+
 v1.1.0 (2022-02-23)
 -------------------
 
-- A new functionnality: a backup file is created each time the user adds a new entry
+- A new functionality: a backup file is created each time the user adds a new entry
 - Better code with a double-linked list
 
 v1.0.1 (2022-01-18)